DX12: delegates update, only one diffuse

+++ DX12/D3D12Research/D3D12/Resources/Shaders/Tonemap/DrawLuminanceHistogram.hlsl
@@ -55,11 +55,11 @@ void DrawLuminanceHistogram(uint groupIndex : SV_GroupIndex, uint3 threadId : SV
 
     // megenta shows the average luminance position
     float4 outColor = BACKGROUND_COLOR;
-    if (floor(tCurrent * NUM_HISTOGRAM_BINS) == groupIndex)
+    if ((uint)floor(tCurrent * NUM_HISTOGRAM_BINS) == groupIndex)
     {
         outColor = CURSOR_COLOR;
     }
-    else if (floor(tTarget * NUM_HISTOGRAM_BINS) == groupIndex)
+    else if ((uint)floor(tTarget * NUM_HISTOGRAM_BINS) == groupIndex)
     {
         outColor = TARGET_COLOR;
     }
